requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.124.4
h11==0.16.0
idna==3.11
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyotp==2.9.0
python-dotenv==1.2.1
qrcode==8.2
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
